requirements.txt
streamlit
tensorflow
numpy
Pillow
requests